invoice_actor/requirements.txt
# python==3.9
streamlit
google-generativeai
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
google-ai-generativelanguage